+++ lib/sequel_rails/storage.rb
@@ -24,13 +24,21 @@ module SequelRails
       adapter.close_connections
       adapter.drop
     end
-    
+
     def self.dump_environment(config_or_env, filename)
       adapter_for(config_or_env).dump(filename)
     end
 
     def self.load_environment(config_or_env, filename)
       adapter_for(config_or_env).load(filename)
+    end
+
+    def self.dump_schema_information(config_or_env, filename)
+      adapter_for(config_or_env).dump_schema_information(filename, :sql => false)
+    end
+
+    def self.dump_schema_information_sql(config_or_env, filename)
+      adapter_for(config_or_env).dump_schema_information(filename, :sql => true)
     end
 
     def self.close_all_connections