+++ spec/internal/db/migrate/1273253849_add_twitter_handle_to_users.rb
@@ -1,0 +1,9 @@
+Sequel.migration do
+
+  change do
+    alter_table :users do
+      add_column :twitter_handle, String, text: true
+    end
+  end
+
+end